feat: custome prompt & render response as markdown #48
