+++ package.json
@@ -26,7 +26,9 @@
     "debug": "^4.3.4",
     "electron-squirrel-startup": "^1.0.0",
     "electron-store": "^8.1.0",
+    "highlight.js": "^11.9.0",
     "https-proxy-agent": "^7.0.4",
+    "markdown-it": "^14.1.0",
     "monaco-editor": "^0.46.0",
     "node-fetch": "^2.7.0",
     "openai": "^4.35.0",
